requirements/base.txt
certifi==2018.8.24
Django==2.1.5
django-filter==2.0.0
pytz==2018.5
psycopg2-binary
reportlab
